API/APP/requirements.txt
fastapi 
uvicorn 
mysql-connector-python 
python-dotenv